:truck: api: move acceptance test to src
